+++ core/transaction/create_multisig_test.go
@@ -45,7 +45,8 @@ func TestCreateMultisigTx(t *testing.T) {
 
 	tx := Transaction{
 		Nonce:         1,
-		GasPrice:      big.NewInt(1),
+		GasPrice:      1,
+		ChainID:       types.ChainTestnet,
 		GasCoin:       coin,
 		Type:          TypeCreateMultisig,
 		Data:          encodedData,
@@ -62,7 +63,7 @@ func TestCreateMultisigTx(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	response := RunTx(cState, false, encodedTx, big.NewInt(0), 0, sync.Map{}, big.NewInt(0))
+	response := RunTx(cState, false, encodedTx, big.NewInt(0), 0, sync.Map{}, 0)
 
 	if response.Code != 0 {
 		t.Fatalf("Response code is not 0. Error %s", response.Log)